Parallelize installation
You don't actually need to wait for one file to download to move onto the next, you can make requests for all files at once.

I believe it may be possible for it to fail on larger repos, in that case it'll be most likely due to ratelimiting. Adding a timer to the http failure part to wait a bit and try again may be worth it in that case. I've never run into this issue personally though.